# openapi/quickeval-v1.yaml
openapi: 3.1.0
info:
  title: QuickEval V1 API
  version: 0.1.0
  description: Internal API for the QuickEval Vue SPA.
servers:
  - url: /
...
paths:
...
  /api/v1/dataset-versions/{version_id}/cases:
    parameters:
      - $ref: "#/components/parameters/VersionID"
    get:
      tags: [Datasets]
      operationId: listVersionCases
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
      responses:
        "200":
          description: Version case page.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VersionCasePageResponse" }
    post:
      tags: [Datasets]
      operationId: createVersionCase
      security:
        - sessionCookie: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VersionCaseRequest" }
      responses:
        "201":
          description: Case created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VersionCaseResponse" }
        "409": { $ref: "#/components/responses/Error" }
...
  /api/v1/badcases:
    get:
      tags: [Badcases]
      operationId: listBadcases
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - name: status
          in: query
          schema: { type: string }
        - name: scenario_id
          in: query
          schema: { $ref: "#/components/schemas/UUID" }
        - name: issue_tag_id
          in: query
          schema: { $ref: "#/components/schemas/UUID" }
        - name: keyword
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Active evaluation Badcases.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BadcasePageResponse" }
  /api/v1/badcases/{badcase_id}:
    parameters:
      - $ref: "#/components/parameters/BadcaseID"
    get:
      tags: [Badcases]
      operationId: getBadcase
      responses:
        "200":
          description: Badcase with evaluation trace, evidence, and activity timeline.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BadcaseResponse" }
        "404": { $ref: "#/components/responses/Error" }
  /api/v1/pages/badcases/{badcase_id}:
    parameters:
      - $ref: "#/components/parameters/BadcaseID"
    get:
      tags: [Badcases]
      operationId: getBadcasePage
      responses:
        "200":
          description: Badcase page aggregate for the SPA.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BadcaseResponse" }
        "404": { $ref: "#/components/responses/Error" }
...
components:
  responses:
    Error:
      description: Request failed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  parameters:
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: page_size
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
...
    VersionID:
      name: version_id
      in: path
      required: true
      schema: { $ref: "#/components/schemas/UUID" }
...
    BadcaseID:
      name: badcase_id
      in: path
      required: true
      schema: { $ref: "#/components/schemas/UUID" }
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: quickeval_session
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
  schemas:
    UUID:
      type: string
      format: uuid
    UTCTimestamp:
      type: string
      format: date-time
      examples:
        - 2026-07-27T06:20:00.000Z
...
    VersionCaseRequest:
      type: object
      required: [user_prompt]
      properties:
        name: { type: [string, "null"], maxLength: 200 }
        user_prompt: { type: string, minLength: 1 }
        precondition: { type: [string, "null"] }
        expected_result: { type: [string, "null"] }
        judging_guide: { type: [string, "null"] }
        is_enabled: { type: boolean, default: true }
        tag_ids:
          type: array
          items: { $ref: "#/components/schemas/UUID" }
        expected_lock_version: { type: integer, minimum: 0, default: 0 }
...
    CaseTagSnapshot:
      type: object
      required: [id, name]
      properties:
        id: { $ref: "#/components/schemas/UUID" }
        name: { type: string }
    VersionCase:
      type: object
      required:
        - id
        - dataset_version_id
        - case_key
        - user_prompt
        - sort_order
        - is_enabled
        - lock_version
        - tags
        - created_at
        - updated_at
      properties:
        id: { $ref: "#/components/schemas/UUID" }
        dataset_version_id: { $ref: "#/components/schemas/UUID" }
        case_key: { $ref: "#/components/schemas/UUID" }
        name: { type: [string, "null"] }
        user_prompt: { type: string }
        precondition: { type: [string, "null"] }
        expected_result: { type: [string, "null"] }
        judging_guide: { type: [string, "null"] }
        sort_order: { type: integer, minimum: 0 }
        is_enabled: { type: boolean }
        lock_version: { type: integer, minimum: 0 }
        tags:
          type: array
          items: { $ref: "#/components/schemas/CaseTagSnapshot" }
        created_at: { $ref: "#/components/schemas/UTCTimestamp" }
        updated_at: { $ref: "#/components/schemas/UTCTimestamp" }
...
    VersionCaseResponse:
      allOf:
        - $ref: "#/components/schemas/ResponseEnvelope"
        - type: object
          properties:
            data: { $ref: "#/components/schemas/VersionCase" }
...
    EvaluationEnvironment:
      type: string
      enum: [test, staging, production, other]
...
    Attachment:
      type: object
      required:
        - id
        - original_name
        - media_type
        - file_size
        - sort_order
        - content_url
        - created_by
        - created_at
      properties:
        id: { $ref: "#/components/schemas/UUID" }
        original_name: { type: string }
        media_type: { type: string, enum: [image/png, image/jpeg, image/webp] }
        file_size: { type: integer, minimum: 1, maximum: 10485760 }
        width: { type: [integer, "null"], minimum: 1 }
        height: { type: [integer, "null"], minimum: 1 }
        sort_order: { type: integer, minimum: 0 }
        content_url: { type: string }
        created_by: { $ref: "#/components/schemas/UUID" }
        created_at: { $ref: "#/components/schemas/UTCTimestamp" }
...
    BadcaseActivity:
      type: object
      required: [id, activity_type, actor_id, actor_name, created_at]
      properties:
        id: { $ref: "#/components/schemas/UUID" }
        activity_type: { type: string, enum: [created, reactivated] }
        note: { type: [string, "null"] }
        actor_id: { $ref: "#/components/schemas/UUID" }
        actor_name: { type: string }
        created_at: { $ref: "#/components/schemas/UTCTimestamp" }
    BadcaseEvaluationContext:
      type: object
      required:
        - case_result_id
        - evaluation_run_id
        - evaluator_id
        - evaluator_name
        - dataset_id
        - dataset_name
        - dataset_version_id
        - version_no
        - version_case_id
      properties:
        case_result_id: { $ref: "#/components/schemas/UUID" }
        evaluation_run_id: { $ref: "#/components/schemas/UUID" }
        evaluator_id: { $ref: "#/components/schemas/UUID" }
        evaluator_name: { type: string }
        dataset_id: { $ref: "#/components/schemas/UUID" }
        dataset_name: { type: string }
        dataset_version_id: { $ref: "#/components/schemas/UUID" }
        version_no: { type: integer, minimum: 1 }
        version_case_id: { $ref: "#/components/schemas/UUID" }
        case_name: { type: [string, "null"] }
        user_prompt: { type: [string, "null"] }
        score: { type: [integer, "null"], minimum: 1, maximum: 5 }
        comment: { type: [string, "null"] }
    Badcase:
      type: object
      required:
        - id
        - source_type
        - scenario_id
        - scenario_name
        - evaluation_target_id
        - evaluation_target_name
        - title
        - environment
        - occurred_at
        - status
        - lock_version
        - created_by
        - creator_name
        - created_at
        - updated_at
        - issue_tags
        - attachments
        - activities
      properties:
        id: { $ref: "#/components/schemas/UUID" }
        source_type: { type: string, enum: [evaluation, business] }
        scenario_id: { $ref: "#/components/schemas/UUID" }
        scenario_name: { type: string }
        evaluation_target_id: { $ref: "#/components/schemas/UUID" }
        evaluation_target_name: { type: string }
        title: { type: string }
        description: { type: [string, "null"] }
        agent_response_text: { type: [string, "null"] }
        agent_version: { type: [string, "null"] }
        environment: { $ref: "#/components/schemas/EvaluationEnvironment" }
        occurred_at: { $ref: "#/components/schemas/UTCTimestamp" }
        status: { type: string }
        lock_version: { type: integer, minimum: 0 }
        created_by: { $ref: "#/components/schemas/UUID" }
        creator_name: { type: string }
        created_at: { $ref: "#/components/schemas/UTCTimestamp" }
        updated_at: { $ref: "#/components/schemas/UTCTimestamp" }
        issue_tags:
          type: array
          items: { $ref: "#/components/schemas/CaseTagSnapshot" }
        evaluation:
          oneOf:
            - $ref: "#/components/schemas/BadcaseEvaluationContext"
            - type: "null"
        attachments:
          type: array
          items: { $ref: "#/components/schemas/Attachment" }
        activities:
          type: array
          items: { $ref: "#/components/schemas/BadcaseActivity" }
    BadcaseResponse:
      allOf:
        - $ref: "#/components/schemas/ResponseEnvelope"
        - type: object
          properties:
            data: { $ref: "#/components/schemas/Badcase" }
    BadcasePageResponse:
      allOf:
        - $ref: "#/components/schemas/ResponseEnvelope"
        - type: object
          properties:
            data:
              allOf:
                - $ref: "#/components/schemas/PageBase"
                - type: object
                  properties:
                    items:
                      type: array
                      items: { $ref: "#/components/schemas/Badcase" }
...
    VersionCasePageResponse:
      allOf:
        - $ref: "#/components/schemas/ResponseEnvelope"
        - type: object
          properties:
            data:
              allOf:
                - $ref: "#/components/schemas/PageBase"
                - type: object
                  properties:
                    items:
                      type: array
                      items: { $ref: "#/components/schemas/VersionCase" }
...
    PageBase:
      type: object
      required: [items, page, page_size, total]
      properties:
        items:
          type: array
          items: {}
        page: { type: integer }
        page_size: { type: integer }
        total: { type: integer }
    ResponseEnvelope:
      type: object
      required: [data, meta]
      properties:
        data: {}
        meta: { $ref: "#/components/schemas/ResponseMeta" }
    ResponseMeta:
      type: object
      required:
        - request_id
      properties:
        request_id:
          type: string
          maxLength: 64
...
    FieldError:
      type: object
      required:
        - field
        - message
      properties:
        field:
          type: string
        message:
          type: string
    Error:
      type: object
      required:
        - code
        - message
        - field_errors
      properties:
        code:
          type: string
        message:
          type: string
        field_errors:
          type: array
          items:
            $ref: "#/components/schemas/FieldError"
        details:
          type: object
          additionalProperties: true
    ErrorResponse:
      type: object
      required:
        - error
        - meta
      properties:
        error:
          $ref: "#/components/schemas/Error"
        meta:
          $ref: "#/components/schemas/ResponseMeta"